automation: Refactor the deploy script and change oper orders
DNS instability still exists when running the test script in CI.
It has been observed that the docker container changes the resolv.conf
when additional networks are added. Therefore, this patch changes the
steps taken in the deploy script such that nmstate is installed from
source *before* the additional networks are added.

In addition, for modularity and ease of reading, a refactoring has been
applied, mainly by extracting commands into functions.

Signed-off-by: Edward Haas <edwardh@redhat.com>